// frontend/wixblog/src/app/shared/services/models/post-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PostDto {
  allowComments?: boolean;
  authorId?: number;
  authorName?: string;
  authorProfilePicture?: string;
  commentCount?: number;
  content?: string;
  createdAt?: string;
  featuredImage?: string;
  id?: number;
  isFeatured?: boolean;
  likeCount?: number;
  likedByCurrentUser?: boolean;
  publishedAt?: string;
  slug?: string;
  status?: 'DRAFT' | 'PUBLISHED' | 'ARCHIVED' | 'DELETED';
  summary?: string;
  timeToRead?: number;
  title?: string;
  updatedAt?: string;
  viewCount?: number;
  viewedByCurrentUser?: boolean;
  wordCount?: number;
}